requirements.txt
langchain
python-dotenv
requests
streamlit
anthropic